rename functions to match code convention
