+++ viziphant/tests/create_targets/create_target_unitary_event_analysis_plot.py
@@ -1,6 +1,6 @@
 import os
 import numpy as np
-from viziphant.unitary_event_analysis_PLOT import \
+from viziphant.unitary_event_analysis_plot import \
     plot_unitary_event_full_analysis, plot_spike_events, plot_spike_rates, \
     plot_coincidence_events, plot_coincidence_rates, \
     plot_statistical_significance, plot_unitary_events
